pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "council-of-elders"
version = "0.1.0"
description = "A local AI advisory system embodying the wisdom of great thinkers"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Council of Elders Contributors" }
]
keywords = ["ai", "wisdom", "ollama", "cli", "advisory"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "ollama>=0.4.0",
    "pyyaml>=6.0",
    "prompt-toolkit>=3.0.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
cloud = [
    "anthropic>=0.40.0",
]
rag = [
    "chromadb>=0.4.0",
]
tts = [
    "elevenlabs>=1.0.0",
]
kindle = [
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
council = "council.cli:app"

[project.urls]
Homepage = "https://github.com/council-of-elders/council"
Documentation = "https://github.com/council-of-elders/council#readme"
Repository = "https://github.com/council-of-elders/council"

[tool.hatch.build.targets.wheel]
packages = ["council"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
